+++ src/main/scala/com/martomate/hexacraft/world/render/BlockRenderer.scala
@@ -47,9 +47,9 @@ class BlockRenderer(val side: Int, init_maxInstances: Int) {
         val x = Math.cos(v).toFloat
         val z = Math.sin(v).toFloat
         BlockVertexData(
-          new Vector3f(x, 1 - s, z),
+          new Vector3f(x, 1f - s, z),
           new Vector2f((1 + (if (s == 0) -x else x)) / 2, (1 + z) / 2),
-          new Vector3f(0, 1 - 2 * s, 0),
+          new Vector3f(0, 1f - 2f * s, 0),
           i
         )
       })
@@ -63,8 +63,8 @@ class BlockRenderer(val side: Int, init_maxInstances: Int) {
         val x = Math.cos(v).toFloat
         val z = Math.sin(v).toFloat
         BlockVertexData(
-          new Vector3f(x, 1 - i / 2, z),
-          new Vector2f(1 - i % 2, i / 2),
+          new Vector3f(x, (1 - i / 2).toFloat, z),
+          new Vector2f((1 - i % 2).toFloat, (i / 2).toFloat),
           new Vector3f(nx, 0, nz),
           i
         )
